Refactors hotkey handling to be scalable
