+++ aas-web-ui/src/App.vue
@@ -19,23 +19,22 @@
 </template>
 
 <script lang="ts" setup>
-    import { computed, onBeforeUnmount, onMounted, ref } from 'vue';
-    import { useRoute, useRouter } from 'vue-router';
+    import { onBeforeUnmount, onMounted, ref } from 'vue';
+    import { useRouter } from 'vue-router';
     import { usePopupOverlay } from '@/composables/PopupOverlay';
-    import { useShortcutManager } from '@/composables/useShortcutManager';
+    import { useGlobalShortcuts } from '@/composables/shortcuts/useGlobalShortcuts';
     import { useInfrastructureStore } from '@/store/InfrastructureStore';
     import { useNavigationStore } from '@/store/NavigationStore';
 
     // Vue Router
     const router = useRouter();
-    const route = useRoute();
 
     // Stores
     const infrastructureStore = useInfrastructureStore();
     const navigationStore = useNavigationStore();
 
-    // Shortcut Manager
-    const shortcuts = useShortcutManager();
+    // Register global shortcuts
+    useGlobalShortcuts();
 
     // Popup Overlay
     const { isPopupOverlayVisible } = usePopupOverlay();
@@ -44,10 +43,6 @@
     const mediaQueryList = window.matchMedia('(max-width: 600px)');
     const matchesMobile = ref(mediaQueryList.matches);
     let tokenRefreshInterval: ReturnType<typeof setInterval> | null = null;
-
-    let unregisterGlobalShortcuts: (() => void) | null = null;
-
-    const isMac = computed(() => typeof navigator !== 'undefined' && /macintosh|mac os x/i.test(navigator.userAgent));
 
     onMounted(() => {
         // Listen for viewport changes (mobile/desktop)
@@ -60,9 +55,6 @@
 
         // Initial token refresh check
         refreshTokens();
-
-        // Register global shortcuts once
-        unregisterGlobalShortcuts = shortcuts.register('global', globalShortcuts /*, { allowInInputs: false } */);
     });
 
     onBeforeUnmount(() => {
@@ -73,9 +65,6 @@
             clearInterval(tokenRefreshInterval);
             tokenRefreshInterval = null;
         }
-
-        unregisterGlobalShortcuts?.();
-        unregisterGlobalShortcuts = null;
     });
 
     /**
@@ -108,62 +97,5 @@
         if (matchesMobile.value !== event.matches) {
             router.go(0); // Reloads current route
         }
-    }
-
-    function isSlashShortcut(e: KeyboardEvent): boolean {
-        // Allow Shift because some layouts need Shift to produce "/"
-        if (e.metaKey || e.ctrlKey || e.altKey) return false;
-
-        return (
-            e.key === '/' || // character-based (works on DE Mac via Shift+7)
-            e.code === 'Slash' || // physical slash key (US layout)
-            e.code === 'NumpadDivide' // numpad /
-        );
-    }
-
-    function globalShortcuts(event: KeyboardEvent): boolean {
-        const keyLower = event.key.toLowerCase();
-
-        // Home: Cmd+Shift+H (Mac), Ctrl+Shift+H (Win/Linux)
-        const isHomeCombo =
-            (isMac.value && event.metaKey && event.shiftKey && !event.ctrlKey && !event.altKey && keyLower === 'h') ||
-            (!isMac.value && event.ctrlKey && event.shiftKey && !event.metaKey && !event.altKey && keyLower === 'h');
-
-        // Cmd/Ctrl+K -> command palette
-        const isCmdK =
-            (isMac.value && event.metaKey && !event.ctrlKey && !event.altKey && !event.shiftKey && keyLower === 'k') ||
-            (!isMac.value && event.ctrlKey && !event.metaKey && !event.altKey && !event.shiftKey && keyLower === 'k');
-
-        // "/" -> navigation palette
-        const isSlash = isSlashShortcut(event);
-
-        if (isHomeCombo) {
-            event.preventDefault();
-            event.stopPropagation();
-
-            router.push({ name: (route.name as string) ?? undefined, query: {} });
-
-            return true;
-        }
-
-        if (isSlash) {
-            event.preventDefault();
-            event.stopPropagation();
-
-            console.log('Navigation palette shortcut "/" triggered');
-            // later: openPalette({ prefill: '/', mode: 'nav' })
-            return true;
-        }
-
-        if (isCmdK) {
-            event.preventDefault();
-            event.stopPropagation();
-
-            console.log('Command palette shortcut Cmd/Ctrl+K triggered');
-            // later: openPalette({ prefill: '>', mode: 'cmd' })
-            return true;
-        }
-
-        return false;
     }
 </script>
