+++ aas-web-ui/src/composables/shortcuts/useShortcutManager.ts
